// src/app/entities/usuario-interface.ts
export interface UsuarioInterface {
    aPaterno?: string;
    aMaterno?: string;
    nombre?: string;
    codigo?: string;
    ESTADO?: number;
    NOMBRE_ESTADO?: string;
    SELECCION?: boolean;
    DNI?: string;
    dependencia?: string;
    unidad?: string;
    id_ugp?: number;
    trama_xml?:string;
    trama_usuario?:string;
  }